test(rules-lint-inert-scoping): merge US-002 lint tests into rules-lint.test.ts
Spec-review flagged that test/unit/cli/rules-lint-inert-scoping.test.ts
violated docs/specs/SPEC-rules-lint-inert-scoping.md's explicit file-placement
directive and the project's one-test-file-per-source-file convention. Merge
its describe blocks into the canonical test/unit/cli/rules-lint.test.ts and
delete the standalone file. Also drops unnecessary type casts around
discoverWorkspacePackages stubbing and a dead _rulesLintDeps save/restore
block that no test in the merged section exercised.
